added recently used exceptions
